% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en_tw_sentences}
\alias{en_tw_sentences}
\title{Sample of English sentences from twitter.}
\format{
An object of class \code{character} of length 1164211.
}
\source{
\url{https://d396qusza40orc.cloudfront.net/dsscapstone/dataset/Coursera-SwiftKey.zip}

For the constructing of this package, a fraction of the source data is randomly selected
with the script:

'data-raw/sampling.r'

as the original files would be too large to be uploaded on GitHub.
The lines in the resulting file in data-raw/en_tw.txt' are cleaned with the function clean_text()
(as included in this package) and made available as the dataset "en_tw_sentences".

To show the first sentence of the dataset:
}
\usage{
en_tw_sentences
}
\description{
Sample of English sentences from twitter.
}
\examples{
en_tw_sentences[1]
}
\keyword{datasets}
