% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscnt.r
\docType{package}
\name{enscnt}
\alias{enscnt}
\title{enscnt: package with dataset of English sentences.}
\description{
Sentences are from ttwitter, blogs and newspapers.
}
\section{functions}{

Additional functions exported:
\code{\link{clean_text}}
}

