% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensentences}
\alias{ensentences}
\title{Sample of English sentences from twitter, blogs and newspapers.}
\format{
An object of class \code{character} of length 823650.
}
\source{
\url{https://d396qusza40orc.cloudfront.net/dsscapstone/dataset/Coursera-SwiftKey.zip}

For the constructing of this package, a fraction of the source data is randomly selected
with the script:

'data-raw/sampling.r'

as the original files would be too large to be uploaded on GitHub.
The lines in the resulting file in data-raw/en.txt' are cleaned with the function clean_text()
(as included in this package) and made available as the dataset "ensentences".

To show the first sentence of the dataset:
}
\usage{
ensentences
}
\description{
Sample of English sentences from twitter, blogs and newspapers.
}
\examples{
ensentences[1]
}
\keyword{datasets}
