% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngrams_table}
\alias{ngrams_table}
\title{Create a dataframe of n-grams, and their frequency and proportion.}
\usage{
ngrams_table(text, n = 2)
}
\arguments{
\item{text}{Previously cleaned text (with \code{enscnt::clean_text()})
(character vector)}

\item{n}{value of 'n' in n-gram (integer)}
}
\value{
dataframe of n-grams (dataframe)
}
\description{
Generate a dataframe of n-grams, and their frequency and proportion
Apply on 'cleaned' text only.
}
