% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngrams_tables}
\alias{ngrams_tables}
\title{Create a list of all n-gram tables 1..ng (wih their frequency and proportion).}
\usage{
ngrams_tables(text, ng = 3, x = NULL)
}
\arguments{
\item{text}{Previously cleaned text (with \code{enscnt::clean_text()})
(character vector)}

\item{ng}{value of 'n' in n-gram (integer)}

\item{x}{list of all n-gram tables (data frames)}
}
\description{
Create a list of all n-gram tables 1..ng (wih their frequency and proportion).
}
\examples{
nGF <- ngrams_tables(en_bl_sentences[1:100], ng=4)
}
