% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampleNlines}
\alias{sampleNlines}
\title{Sample "aFraction" of cleaned text.}
\usage{
sampleNlines(text, aFraction = 0.1)
}
\arguments{
\item{text}{Previously cleaned text (with \code{enscnt::clean_text()})
(character vector)}

\item{aFraction}{Fraction of lines in text to be used \code{0 < aFraction < 1} (numeric)}
}
\value{
Cleaned text (character vector)
}
\description{
Sample "aFraction" of cleaned text.
Apply on 'cleaned' text only.
}
