% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setProbsOfnGrams}
\alias{setProbsOfnGrams}
\title{Estimate the probabilities of observed and unobserved n-grams that start with prefix "pfx".}
\usage{
setProbsOfnGrams(pfx, nGF, discnt = 0.5)
}
\arguments{
\item{pfx}{prefix ('cleaned' text only, character)}

\item{nGF}{List of all n-gram tables 1..ng wih their frequency and proportion
as created by the function ngrams_tables (list)}

\item{discnt}{Discount factor 0 < discnt < 1 (numeric)}
}
\value{
data.frame of 6 variables:
ngrams: unique n-grams (character)
freq: frequency of n-grams (integer)
prop: proportion of n-grams (numeric)
ng: value of 'n' in n-gram (integer)
prob: probability of n-gram (numeric)
}
\description{
It accomplishes this estimation by backing off through progressively shorter history models
under certain conditions. Reference: \href{https://en.wikipedia.org/wiki/Katz\%27s_back-off_model}{Katz's back-off model}.
}
\details{
By doing so, the model with the most reliable information about a given history is used
to provide the better results in predicting the next word.
}
\examples{
nGF <- ngrams_tables(en_bl_sentences[1:100], ng=4)
nGF_prob <- setProbsOfnGrams(pfx="love spending", nGF, discnt=0.75)
}
