% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{word_count}
\alias{word_count}
\title{Count words in cleaned character vector.}
\usage{
word_count(text)
}
\arguments{
\item{text}{Previously cleaned text (with \code{enscnt::clean_text()})
(character vector)}
}
\value{
Number of words (integer)
}
\description{
Count words in cleaned character vector.
}
