% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_text}
\alias{clean_text}
\title{Transform text with non-text characters to 'clean' text}
\usage{
clean_text(x)
}
\arguments{
\item{x}{Text to clean (character).}
}
\value{
Cleaned text (character)
}
\description{
Remove non-text characters. Numbers are also removed.
}
