% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscnt.r
\docType{package}
\name{enscnt}
\alias{enscnt}
\title{enscnt: package with datasets of English sentences.}
\description{
Sentences are from twitter, blogs and newspapers.
}
\section{datasets}{

\code{\link{ensentences}}

\code{\link{en_bl_sentences}}

\code{\link{en_ne_sentences}}

\code{\link{en_tw_sentences}}
}

\section{functions}{

Additional functions exported:

\code{\link{clean_text}}

\code{\link{word_count}}

\code{\link{ngrams_table}}

\code{\link{sampleNlines}}

\code{\link{plot_ngrams}}
}

