% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscnt.r
\docType{package}
\name{enscnt}
\alias{enscnt}
\title{enscnt: package with dataset of English sentences.}
\description{
Sentences are from twitter, blogs and newspapers.
}
\section{dataset}{

\code{\link{ensentences}}
}

\section{functions}{

Additional functions exported:
\code{\link{clean_text}}
\code{\link{word_count}}
\code{\link{word_count}}
\code{\link{word_count}}
\code{\link{word_count}}
\code{\link{word_count}}
}

