% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensentences}
\alias{ensentences}
\title{Sample of English sentences.}
\format{
An object of class \code{character} of length 824119.
}
\source{
\url{https://d396qusza40orc.cloudfront.net/dsscapstone/dataset/Coursera-SwiftKey.zip}

For constructing this package, 25\% of this data is selected with the script
'data-raw/sampling.r' and written in a file 'data-raw/en.txt'.

(The original files would be too large to be uploaded on GitHub).

The lines in 'data-raw/en.txt' are cleaned with the function clean_text()
(as included in this package) and made available as a dataset 'ensentences'.
}
\usage{
ensentences
}
\description{
A dataset containing sample sentences from twitter, blogs and newspapers.
}
\examples{
ensentences[1]
}
\keyword{datasets}
