% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ngrams}
\alias{plot_ngrams}
\title{Plot n-gram (output from \code{ngram::ngram})}
\usage{
plot_ngrams(ngr, title = "", n = 2, m = 5)
}
\arguments{
\item{ngr}{ngram (output from \code{ngram::ngram})}

\item{title}{Plot title (character)}

\item{n}{value of 'n' in n-gram (integer)}

\item{m}{most frequent ngrams (integer)}
}
\value{
(-) Only side-effect is produces (a plot)
}
\description{
Plot n-gram (output from \code{ngram::ngram})
}
